Add an example for searching for any of multiple elements, returning the first one found
The use case for this example is a situation when there are finite defined
scenarios where the transaction may end up in. Finding an element that is
specific to any of the given scenarios determines where the transaction
ended up at, and branching can be performed to handle each specific scenario
correctly.